Better conveyor physics / Performance test maps / Unit inventory begin

+++ core/src/io/anuke/mindustry/resource/AmmoType.java
@@ -53,7 +53,7 @@ public class AmmoType {
         return allTypes;
     }
 
-    public static AmmoType getByID(byte id){
+    public static AmmoType getByID(int id){
         return allTypes.get(id);
     }
 }
